requirements.txt
imgaug==0.4.0
keras==2.11.0
keras_resnet==0.2.0
matplotlib==3.7.1
mlxtend==0.21.0
numpy==1.24.2
opencv_python==4.6.0.66
Pillow==9.5.0
pyclipper==1.3.0.post4
seaborn==0.12.2
shapely==2.0.1
tensorflow==2.12.0
tensorflow_intel==2.11.0
ttkbootstrap==1.10.1
